Was separated index.js file in more files, where each new file will store it corresponding component (each component - one class). Also, the corrections to setState calls were fixed.
