+++ inginious/frontend/webapp/plugins/problem_bank/react_app/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "jquery": "^3.3.1",
     "react": "^16.0.0",
     "react-autosuggest": "^9.3.2",
     "react-bootstrap": "^0.31.5",
